requirements.txt
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
joblib 
kiwisolver==1.4.4
matplotlib==3.7.1
nltk==3.8.1
numpy 
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
scikit-learn==1.2.2
scikit-surprise==1.1.3
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.65.0
wincertstore==0.2
fastapi
uvicorn
pymongo